+++ app/Http/Controllers/Admin/ProductController.php
@@ -6,10 +6,11 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Repositories\ProductRepository;
 use App\Repositories\CateProductRepository;
+use App\Http\Requests\ProductRequest;
+use App\Http\Requests\EditProductRequest;
 use Auth;
 use Hash;
 use File;
-use App\Http\Requests\ProductRequest;
 
 class ProductController extends Controller
 {
@@ -60,20 +61,28 @@ class ProductController extends Controller
         if (!empty($request->file('fImage'))) {
             $file_name = $request->file('fImage')->getClientOriginalName();
             $image = 'uploads/product/' . time() . '-' . $file_name;
-            $request->file('fImage')->move('uploads/news/', $image);
+            $request->file('fImage')->move('uploads/product/', $image);
+        }
+        if ($request->sale != 0) {
+            $price_new = $request->price_old - ($request->price_old * $request->sale) / 100;
+        } else {
+            $price_new = 0;
         }
 
         $request->merge(
             [
                 'user_id' => $user->id,
                 'slug' => str_slug($request->name),
-                'thumbnail' => $image
+                'thumbnail' => !empty($image) == true ? $image : null,
+                'price_new' => $price_new
             ]
         );
-
         $this->product->create($request->all());
 
-        return redirect(route('product.index'));
+        return redirect(route('product.index'))->with([
+            'flash_level' => 'success',
+            'flash_message' => 'Thêm thành công !'
+        ]);
     }
 
     /**
@@ -95,7 +104,10 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
-        //
+        $product = $this->product->findProduct($id);
+        $cate_product = $this->cate_product->listCateProductOpen();
+
+        return view('backend.product.edit', compact('product', 'cate_product'));
     }
 
     /**
@@ -105,9 +117,55 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(EditProductRequest $request, $id)
     {
-        //
+        if (Auth::check()) {
+            $user = Auth::user();
+        }
+        $product = $this->product->findProduct($id);
+
+        if (!empty($request->file('fImage'))) {
+            $image = $product->thumbnail;
+            $file_name      = $request->file('fImage')->getClientOriginalName();
+            $thumbnail    = 'uploads/product/'.time().'-'.$file_name;
+            $request->file('fImage')->move('uploads/product/', $thumbnail);
+            if(File::exists($image)){
+                File::delete($image);
+            }
+        }
+
+        $sale = $product->sale;
+        if ($request->sale == 0)
+        {
+            $price_new = 0;
+        } elseif ($request->sale == $sale){
+            $price_new = $product->price_new;
+        } else
+        {
+            $price_new = $request->price_old - ($request->price_old * $request->sale) / 100;
+        }
+
+        if (empty($thumbnail)){
+            $image_product = $product->thumbnail;
+        } else
+        {
+            $image_product = $thumbnail;
+        }
+
+        $request->merge(
+            [
+                'user_id' => $user->id,
+                'thumbnail' => $image_product,
+                'price_new' => $price_new
+            ]
+        );
+
+        $this->product->update($id, $request->all());
+
+        return redirect()->route('product.index')->with([
+            'flash_level' => 'success',
+            'flash_message' => 'Cập nhật thành công !'
+        ]);
     }
 
     /**
@@ -118,8 +176,17 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $this->product->delete($id);
-
-        return redirect(route('product.index'));
+        $product = $this->product->findProduct($id);
+        if (isset($product)) {
+            $image = $product->thumbnail;
+            if(File::exists($image)){
+                File::delete($image);
+            }
+            $this->product->delete($id);
+        }
+        return back()->with([
+            'flash_level' => 'success',
+            'flash_message' => 'Xóa thành công !'
+        ]);
     }
 }
